feat(implemet other adpter methods and move err decorator to the class)
