+++ src/domain/async_service_api_adapter.py
@@ -1,11 +1,77 @@
+import functools
+import logging as log
 from typing import Optional, Dict
 
 import httpx
 from httpx import Response
-from starlette.responses import JSONResponse
+from starlette import status
 
+from src.config.exception import ClientException, ForbiddenException, UnauthorizedException, NotFoundException, \
+    NotAcceptableException, ServerException
 from src.domain.service_api import IServiceApi
-from src.infra.util.util import check_response_code
+
+
+def check_response_code(method: str, expected_code: int = 200):
+    def decorator_check_response_code(func):
+
+        @functools.wraps(func)
+        async def wrapper_check_response_code(*args, **kwargs):
+            function_name: str = func.__name__
+            url = kwargs.get('url', None)
+            params = kwargs.get('params', None)
+            body = kwargs.get('json', None)
+            headers = kwargs.get('headers', None)
+            err: str = ''
+            try:
+                service_api_res = await func(*args, **kwargs)
+                # 代表是simple系列
+                if type(service_api_res) is not tuple:
+                    return service_api_res
+                data: Dict = service_api_res[0]
+                msg: str = service_api_res[1]
+                err: str = service_api_res[2]
+                # ex: 40400 取至404
+                status_code: int = int(service_api_res[3][:3]) if len(service_api_res) >= 3 else None
+                if status_code is not None and status_code < 400 or status_code == expected_code:
+                    return service_api_res
+
+                log.error(
+                    f"service request fail, [%s]: %s, body:%s, params:%s, headers:%s, status_code:%s, msg:%s, \n response:%s",
+                    method, url, body, params, headers, status_code, msg, service_api_res)
+                if status_code is not None:
+
+                    if status_code == status.HTTP_400_BAD_REQUEST:
+                        raise ClientException(msg=msg, data=data)
+
+                    if status_code == status.HTTP_401_UNAUTHORIZED:
+                        raise UnauthorizedException(msg=msg, data=data)
+
+                    if status_code == status.HTTP_403_FORBIDDEN:
+                        raise ForbiddenException(msg=msg, data=data)
+
+                    if status_code == status.HTTP_404_NOT_FOUND:
+                        raise NotFoundException(msg=msg, data=data)
+
+                    if status_code == status.HTTP_406_NOT_ACCEPTABLE:
+                        raise NotAcceptableException(msg=msg, data=data)
+
+                raise ServerException(msg=msg, data=data)
+            except Exception as e:  # 這邊捕捉Server side err
+                err = e.__str__()
+
+                err_msg: str = (
+                    f"unhandled {function_name} request error,\n"
+                    f"url:{url}, \n"
+                    f"params: {params}\n, "
+                    f"headers: {headers}\n,"
+                    f"error detail: {err} \n,"
+                )
+                log.error(err_msg)
+                raise ServerException(msg=err_msg)
+
+        return wrapper_check_response_code
+
+    return decorator_check_response_code
 
 
 class AsyncServiceApiAdapter(IServiceApi):
@@ -40,7 +106,7 @@ class AsyncServiceApiAdapter(IServiceApi):
         dto: Dict = {}  # init with empty dict
         msg: str = ''
         err: str = ''
-        code: int = None
+        code: Optional[int] = None
         try:
             async with httpx.AsyncClient() as client:
                 response: Response = await client.get(url=url, headers=headers)
@@ -53,84 +119,135 @@ class AsyncServiceApiAdapter(IServiceApi):
 
         return dto.get('data'), msg, err, code  # 這裡應該是對應service的VO/responseDTO+msg
 
-    async def simple_post(self, url: str, json: Dict, headers: Dict = None) -> (Optional[Dict[str, str]]):
+    @check_response_code('post', 200)
+    async def simple_post(self, url: str, data: Dict = None, headers: Dict = None) -> Optional[Dict[str, str]]:
         async with httpx.AsyncClient() as client:
-            response: Response = await client.post(url=url, json=json, headers=headers)
-            dto: Dict = response.json()
-            return dto.get('data')  # 這裡應該是對應service的VO/responseDTO
+            response: Response = await client.post(url=url, json=data, headers=headers)
 
+            dto: Dict = response.json()
+
+            return dto.get('data')
+
+    @check_response_code('post', 200)
+    async def post(self, url: str, data: Dict = None, headers: Dict = None) -> (
+            Optional[Dict[str, str]], Optional[str], Optional[str]):
+        dto: Dict = {}  # init with empty dict
+        msg: str = ''
+        err: str = ''
+
+        async with httpx.AsyncClient() as client:
+            response: Response = await client.post(url=url, json=data, headers=headers)
+
+            dto: Dict = response.json()
+            msg: str = dto.get('msg')
+
+            return dto.get('data'), msg, err
+
+    @check_response_code('post', 200)
+    async def post_with_statuscode(self, url: str, data: Dict = None, headers: Dict = None) -> (
+            Optional[Dict[str, str]], Optional[str], Optional[int], Optional[str]):
+        dto: Dict = {}  # init with empty dict
+        msg: str = ''
+        err: str = ''
+        code: Optional[int] = None
+        try:
+            async with httpx.AsyncClient() as client:
+                response: Response = await client.post(url=url, json=data, headers=headers)
+
+                dto: Dict = response.json()
+                msg: Optional[str] = dto.get('msg', None)
+                code: Optional[int] = dto.get('code', None)
+        except Exception as e:
+            err = e.__str__()
+
+        return dto.get('data'), msg, err, code
+
+    @check_response_code('post', 200)
     async def post_data(self, url: str, byte_data: bytes, headers: Dict = None) -> (Optional[Dict[str, str]]):
         pass
 
-    async def post(self, url: str, json: Dict, headers: Dict = None) -> (
+    @check_response_code('put', 200)
+    async def simple_put(self, url: str, data: Dict = None, headers: Dict = None) -> Optional[Dict[str, str]]:
+        async with httpx.AsyncClient() as client:
+            response: Response = await client.put(url=url, json=data, headers=headers)
+
+            dto: Dict = response.json()
+
+            return dto.get('data')
+
+    @check_response_code('put', 200)
+    async def put(self, url: str, data: Dict = None, headers: Dict = None) -> (
             Optional[Dict[str, str]], Optional[str], Optional[str]):
+        dto: Dict = {}  # init with empty dict
+        msg: str = ''
+        err: str = ''
+
         async with httpx.AsyncClient() as client:
-            response: Response = await client.post(url=url, json=json, headers=headers)
+            response: Response = await client.put(url=url, json=data, headers=headers)
+
             dto: Dict = response.json()
-            server_res: JSONResponse = JSONResponse(content={
-                'msg': dto.get('msg'),
-                'data': dto.get('data')
-            })
+            msg: str = dto.get('msg')
 
-            return server_res  # 這裡應該是對應service的VO/responseDTO+msg
+            return dto.get('data'), msg, err
 
-    async def post_with_statuscode(self, url: str, json: Dict, headers: Dict = None) -> (
+    @check_response_code('put', 200)
+    async def put_with_statuscode(self, url: str, data: Dict = None, headers: Dict = None) -> (
             Optional[Dict[str, str]], Optional[str], Optional[int], Optional[str]):
-        async with httpx.AsyncClient() as client:
-            response: Response = await client.post(url=url, json=json, headers=headers)
-            # 直接回傳服務的Response
-            result = response.json()
-            status_code = response.status_code
-            msg = result["msg"]
-            result = result["data"]
-            return result, msg, status_code, err
+        dto: Dict = {}  # init with empty dict
+        msg: str = ''
+        err: str = ''
+        code: Optional[int] = None
+        try:
+            async with httpx.AsyncClient() as client:
+                response: Response = await client.put(url=url, json=data, headers=headers)
 
-    async def simple_put(self, url: str, json: Dict = None, headers: Dict = None) -> (Optional[Dict[str, str]]):
+                dto: Dict = response.json()
+                msg: Optional[str] = dto.get('msg', None)
+                code: Optional[int] = dto.get('code', None)
+        except Exception as e:
+            err = e.__str__()
+
+        return dto.get('data'), msg, err, code
+
+    @check_response_code('delete', 200)
+    async def simple_delete(self, url: str, params: Dict = None, headers: Dict = None) -> Optional[Dict[str, str]]:
         async with httpx.AsyncClient() as client:
-            response: Response = await client.post(url=url, json=json, headers=headers)
+            response: Response = await client.delete(url=url, headers=headers, params=params)
+
             dto: Dict = response.json()
-            return dto.get('data')  # 這裡應該是對應service的VO/responseDTO
 
-    async def put(self, url: str, json: Dict = None, headers: Dict = None) -> (
-            Optional[Dict[str, str]], Optional[str], Optional[str]):
-        async with httpx.AsyncClient() as client:
-            response: Response = await client.put(url=url, json=json, headers=headers)
-            dto: Dict = response.json()
-            server_res: JSONResponse = JSONResponse(content={
-                'msg': dto.get('msg'),
-                'data': dto.get('data')
-            })
+            return dto.get('data')
 
-            return server_res  # 這裡應該是對應service的VO/responseDTO+msg
-
-    async def put_with_statuscode(self, url: str, json: Dict = None, headers: Dict = None) -> (
-            Optional[Dict[str, str]], Optional[str], Optional[int], Optional[str]):
-        async with httpx.AsyncClient() as client:
-            response: Response = await client.put(url=url, json=json, headers=headers)
-            # 直接回傳服務的Response
-            return response.json()
-
-    async def simple_delete(self, url: str, params: Dict = None, headers: Dict = None) -> (Optional[Dict[str, str]]):
-        async with httpx.AsyncClient() as client:
-            response: Response = await client.delete(url=url, headers=headers)
-            dto: Dict = response.json()
-            return dto.get('data')  # 這裡應該是對應service的VO/responseDTO
-
+    @check_response_code('delete', 200)
     async def delete(self, url: str, params: Dict = None, headers: Dict = None) -> (
             Optional[Dict[str, str]], Optional[str], Optional[str]):
+        dto: Dict = {}  # init with empty dict
+        msg: str = ''
+        err: str = ''
+
         async with httpx.AsyncClient() as client:
-            response: Response = await client.delete(url=url, headers=headers)
+            response: Response = await client.delete(url=url, headers=headers, params=params)
+
             dto: Dict = response.json()
-            server_res: JSONResponse = JSONResponse(content={
-                'msg': dto.get('msg'),
-                'data': dto.get('data')
-            })
+            msg: str = dto.get('msg')
 
-            return server_res  # 這裡應該是對應service的VO/responseDTO+msg
+            return dto.get('data'), msg, err
 
+    @check_response_code('delete', 200)
     async def delete_with_statuscode(self, url: str, params: Dict = None, headers: Dict = None) -> (
             Optional[Dict[str, str]], Optional[str], Optional[int], Optional[str]):
-        async with httpx.AsyncClient() as client:
-            response: Response = await client.delete(url=url, headers=headers)
-            # 直接回傳服務的Response
-            return response.json()
+        dto: Dict = {}  # init with empty dict
+        msg: str = ''
+        err: str = ''
+        code: Optional[int] = None
+        try:
+            async with httpx.AsyncClient() as client:
+                response: Response = await client.delete(url=url, headers=headers, params=params)
+
+                dto: Dict = response.json()
+                msg: Optional[str] = dto.get('msg', None)
+                code: Optional[int] = dto.get('code', None)
+        except Exception as e:
+            err = e.__str__()
+
+        return dto.get('data'), msg, err, code
